PIM-3595: move detacher to common folder
