+++ spec/Akeneo/Bundle/StorageUtilsBundle/Doctrine/Common/Detacher/ObjectDetacherSpec.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace spec\Akeneo\Bundle\StorageUtilsBundle\Doctrine;
+namespace spec\Akeneo\Bundle\StorageUtilsBundle\Doctrine\Common\Detacher;
 
 use Doctrine\Common\Persistence\ManagerRegistry;
 use Doctrine\Common\Persistence\ObjectManager;
