+++ src/Akeneo/Bundle/StorageUtilsBundle/Resources/config/doctrine.yml
@@ -3,7 +3,7 @@ parameters:
     doctrine.document_repository.class: Doctrine\ODM\MongoDB\DocumentRepository
     akeneo_storage_utils.doctrine.smart_manager_registry.class: Akeneo\Bundle\StorageUtilsBundle\Doctrine\SmartManagerRegistry
     akeneo_storage_utils.doctrine.table_name_builder.class:     Akeneo\Bundle\StorageUtilsBundle\Doctrine\TableNameBuilder
-    akeneo_storage_utils.doctrine.object_detacher.class:        Akeneo\Bundle\StorageUtilsBundle\Doctrine\ObjectDetacher
+    akeneo_storage_utils.doctrine.object_detacher.class:        Akeneo\Bundle\StorageUtilsBundle\Doctrine\Common\Detacher\ObjectDetacher
     akeneo_storage_utils.event_subscriber.resolve_target_repository.class: Akeneo\Bundle\StorageUtilsBundle\EventSubscriber\ResolveTargetRepositorySubscriber
 
 services:
